stringify for telemetry new nested data part for bare uint8arrays
